+++ src/UdpSocket.h
@@ -29,7 +29,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 class UdpSocket : public Socket
 {
 public:
-	UdpSocket(SocketHandler& ,size_t ibufsz = 16384);
+	UdpSocket(SocketHandler& ,int ibufsz = 16384);
 	~UdpSocket();
 
 	/** new callback */
@@ -46,13 +46,13 @@ public:
 	bool Open6(const std::string& host,port_t port);
 
 	/** send to specified address */
-	void SendToBuf(const std::string& ,port_t,const char *data,size_t len,int flags = 0);
-	void SendToBuf6(const std::string& ,port_t,const char *data,size_t len,int flags = 0);
+	void SendToBuf(const std::string& ,port_t,const char *data,int len,int flags = 0);
+	void SendToBuf6(const std::string& ,port_t,const char *data,int len,int flags = 0);
 	void SendTo(const std::string&,port_t,const std::string&,int flags = 0);
 	void SendTo6(const std::string&,port_t,const std::string&,int flags = 0);
 
 	/** send to connected address */
-	void SendBuf(const char *data,size_t,int flags = 0);
+	void SendBuf(const char *data,int,int flags = 0);
 	void Send(const std::string& ,int flags = 0);
 
 	/** broadcast */
@@ -74,13 +74,13 @@ protected:
 	UdpSocket(const UdpSocket& s) : Socket(s) {}
 	void OnRead();
 
-// int  recvfrom(int  s,  void *buf, size_t len, int flags, struct sockaddr *from, socklen_t *fromlen);
-// int  sendto(int  s,  const  void  *msg, size_t len, int flags, const struct sockaddr *to, socklen_t tolen);
+// int  recvfrom(int  s,  void *buf, int len, int flags, struct sockaddr *from, socklen_t *fromlen);
+// int  sendto(int  s,  const  void  *msg, int len, int flags, const struct sockaddr *to, socklen_t tolen);
 private:
 	UdpSocket& operator=(const UdpSocket& ) { return *this; }
 	bool m_connected;
 	char *m_ibuf;
-	size_t m_ibufsz;
+	int m_ibufsz;
 };
